% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DensityPrediction.R
\name{as_prediction_density}
\alias{as_prediction_density}
\alias{as_prediction_density.PredictionDensity}
\alias{as_prediction_density.data.frame}
\title{Convert to a Density Estimation Prediction}
\usage{
as_prediction_density(x, ...)

\method{as_prediction_density}{PredictionDensity}(x, ...)

\method{as_prediction_density}{data.frame}(x, ...)
}
\value{
\code{\link{PredictionDensity}}.
}
\description{
Convert an object to a \code{\link{PredictionDensity}}
}
\seealso{
Other density estimation classes: 
\code{\link{LearnerDensity}},
\code{\link{MeasureDensity}},
\code{\link{PredictionDensity}},
\code{\link{TaskDensity}},
\code{\link{as_task_density}()},
\code{\link{mlr_learners_density.featureless}},
\code{\link{mlr_learners_density.np}},
\code{\link{mlr_measures_density.logloss}},
\code{\link{mlr_tasks_faithful}}
}
\concept{density estimation classes}
