% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DensityTask.R
\name{as_task_density}
\alias{as_task_density}
\alias{as_task_density.TaskDensity}
\alias{as_task_density.data.frame}
\alias{as_task_density.DataBackend}
\title{Convert to a Density Task}
\usage{
as_task_density(x, ...)

\method{as_task_density}{TaskDensity}(x, clone = FALSE, ...)

\method{as_task_density}{data.frame}(x, id = deparse1(substitute(x)), ...)

\method{as_task_density}{DataBackend}(x, id = deparse1(substitute(x)), ...)
}
\arguments{
\item{x}{(any)\cr
Object to convert.}

\item{...}{(any)\cr
Additional arguments.}

\item{clone}{(\code{logical(1)})'cr
Whether to ensure the returned object is different from input \code{x}.}

\item{id}{(\code{character(1)})\cr
ID for the new \code{\link[mlr3:Task]{Task}}.
Defaults to the (deparsed) name of the expression given to \code{x}.}
}
\value{
\code{\link{TaskDensity}}
}
\description{
Convert to a \code{\link{TaskDensity}}.
}
\seealso{
Other density estimation classes: 
\code{\link{LearnerDensity}},
\code{\link{MeasureDensity}},
\code{\link{PredictionDensity}},
\code{\link{TaskDensity}},
\code{\link{as_prediction_density}()},
\code{\link{mlr_learners_density.featureless}},
\code{\link{mlr_learners_density.np}},
\code{\link{mlr_measures_density.logloss}},
\code{\link{mlr_tasks_faithful}}
}
\concept{density estimation classes}
