% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionaries.R
\docType{data}
\name{dict_filtors}
\alias{dict_filtors}
\title{Dictionary of Filtors}
\format{
An object of class \code{DictionaryFiltor} (inherits from \code{Dictionary}, \code{R6}) of length 13.
}
\usage{
dict_filtors
}
\description{
Dictionary of Filtors
}
\keyword{datasets}
