% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionaries.R
\docType{data}
\name{dict_mutators}
\alias{dict_mutators}
\title{Dictionary of Mutators}
\format{
An object of class \code{DictionaryMutator} (inherits from \code{Dictionary}, \code{R6}) of length 13.
}
\usage{
dict_mutators
}
\description{
Dictionary of Mutators
}
\keyword{datasets}
