% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mutator.R
\name{dict_mutators_maybe}
\alias{dict_mutators_maybe}
\alias{MutatorMaybe}
\title{Mutator Choosing Action Probabilistically}
\description{
\code{\link{Mutator}} that chooses which operation to perform probabilistically. The \code{\link{Mutator}} wraps two other \code{\link{Mutator}}s given during construction,
and for each individuum, the operation to perform is sampled: with probability \code{p} (hyperparameter), the \code{\link{Mutator}} given to the \code{mutator}
construction argument is applied, and with probability \code{p - 1} the one given to \code{mutator_not} is applied.
}
\section{Hyperparameters}{

This operator has the hyperparameters of the \code{\link{Mutator}}s that it wraps: The hyperparameters of the operator given to the \code{mutator} construction argument
are prefixed with \code{"maybe."}, the hyperparameters of the operator given to the \code{mutator_not} construction argument are prefixed with \code{"maybe_not."}.

Additional hyperparameters:
\itemize{
\item \code{p} :: \code{numeric(1)} \cr
Probability per individual with which to apply the operator given to the \code{mutator} construction argument.
}
}

\section{Supported Operand Types}{


Supported \code{\link[paradox:Param]{Param}} classes are the set intersection of supported classes of \code{mutator} and \code{mutator_not}.
}

\section{Dictionary}{

This \code{\link{Mutator}} can be created with the short access form \code{\link[=mut]{mut()}}
(\code{\link[=muts]{muts()}} to get a list), or through the the \link[mlr3misc:Dictionary]{dictionary}
\code{\link{dict_mutators}} in the following way:\preformatted{# preferred:
mut("maybe", <mutator> [, <mutator_not>])
muts("maybe", <mutator> [, <mutator_not>])  # takes vector IDs, returns list of [`Mutator`]s

# long form:
dict_mutators$get("maybe", <mutator> [, <mutator_not>])
}
}

\seealso{
Other mutators: 
\code{\link{MutatorDiscrete}},
\code{\link{MutatorNumeric}},
\code{\link{MutatorProxy}},
\code{\link{Mutator}},
\code{\link{OperatorCombination}},
\code{\link{dict_mutators_cmpmaybe}},
\code{\link{dict_mutators_gauss}},
\code{\link{dict_mutators_null}},
\code{\link{dict_mutators_unif}}

Other mutator wrappers: 
\code{\link{MutatorProxy}},
\code{\link{OperatorCombination}},
\code{\link{dict_mutators_cmpmaybe}}
}
\concept{mutator wrappers}
\concept{mutators}
\section{Super classes}{
\code{\link[miesmuschel:MiesOperator]{miesmuschel::MiesOperator}} -> \code{\link[miesmuschel:Mutator]{miesmuschel::Mutator}} -> \code{MutatorMaybe}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{MutatorMaybe$new()}}
\item \href{#method-prime}{\code{MutatorMaybe$prime()}}
\item \href{#method-clone}{\code{MutatorMaybe$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="miesmuschel" data-topic="MiesOperator" data-id="operate">}\href{../../miesmuschel/html/MiesOperator.html#method-operate}{\code{miesmuschel::MiesOperator$operate()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initialize the \code{MutatorMaybe} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MutatorMaybe$new(mutator, mutator_not = MutatorNull$new())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{mutator}}{(\code{\link{Mutator}})\cr
\code{\link{Mutator}} to wrap. This operator gets run with probability \code{p} (Hyperparameter).\cr
The constructed object gets a \emph{clone} of this argument.}

\item{\code{mutator_not}}{(\code{\link{Mutator}})\cr
Another \code{\link{Mutator}} to wrap. This operator runs when \code{mutator} is not chosen. By
default, this is \code{\link{MutatorNull}}, i.e. no operation. With this default, the
\code{MutatorMaybe} object applies the \code{mutator} operation with probability \code{p}, and
no operation at all otherwise.\cr
The constructed object gets a \emph{clone} of this argument.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-prime"></a>}}
\if{latex}{\out{\hypertarget{method-prime}{}}}
\subsection{Method \code{prime()}}{
See \code{\link{MiesOperator}} method. Primes both this operator, as well as the wrapped operators
given to \code{mutator} and \code{mutator_not} during construction.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MutatorMaybe$prime(param_set)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{param_set}}{(\code{\link[paradox:ParamSet]{ParamSet}})\cr
Passed to \code{\link{MiesOperator}}\verb{$prime()}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\link{invisible} \code{self}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MutatorMaybe$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
