% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MutatorNumericPolynomial.R
\name{dict_mutators_numericpoly}
\alias{dict_mutators_numericpoly}
\alias{MutatorNumericPolynomial}
\title{Polynomial Numeric Mutator}
\description{
Numeric components are mutated by sampling from a polynomial distribution having its mean at the
current value and its variance as a function of the distribution index \code{n}.
See Deb (1996) for more details.

Since the information loss is rather high, this should in most cases be combined with \code{\link{MutatorCmpMaybe}}.
}
\section{Configuration Parameters}{

\itemize{
\item \code{n} :: \code{double}\cr
Non-negative distribution index of the polynomial distribution for each individual.
Generally spoken, the higher \code{n}, the lower the variance of the distribution.
This may either be a scalar in which case it is applied to all input components, or a vector,
in which case it must have the length of the input and applies to components in order in which
they appear in the priming \code{\link[paradox:ParamSet]{ParamSet}}. Initialized to 1.
}
}

\section{Supported Operand Types}{


Supported \code{\link[paradox:Param]{Param}} classes are: \code{\link[paradox:ParamInt]{ParamInt}}, \code{\link[paradox:ParamDbl]{ParamDbl}}
}

\section{Dictionary}{

This \code{\link{Mutator}} can be created with the short access form \code{\link[=mut]{mut()}}
(\code{\link[=muts]{muts()}} to get a list), or through the the \link[mlr3misc:Dictionary]{dictionary}
\code{\link{dict_mutators}} in the following way:\preformatted{# preferred:
mut("numericpoly")
muts("numericpoly")  # takes vector IDs, returns list of Mutators

# long form:
dict_mutators$get("numericpoly")
}
}

\examples{
set.seed(1)
mnp = mut("numericpoly")
p = ps(x = p_int(-5, 5), y = p_dbl(-5, 5))
data = data.frame(x = rep(0, 1000), y = rep(0, 1000))

mnp$prime(p)
mnp$operate(data)
}
\references{
Deb, Kalyanmoy, Goyal, Mayank (1996).
\dQuote{A combined genetic adaptive search (GeneAS) for engineering design.}
\emph{Computer Science and Informatics}, \bold{26}(4), 30--45.
}
\seealso{
Other mutators: 
\code{\link{MutatorDiscrete}},
\code{\link{MutatorNumeric}},
\code{\link{MutatorProxy}},
\code{\link{Mutator}},
\code{\link{OperatorCombination}},
\code{\link{dict_mutators_cmpmaybe}},
\code{\link{dict_mutators_gauss}},
\code{\link{dict_mutators_maybe}},
\code{\link{dict_mutators_null}},
\code{\link{dict_mutators_unif}}
}
\concept{mutators}
\section{Super classes}{
\code{\link[miesmuschel:MiesOperator]{miesmuschel::MiesOperator}} -> \code{\link[miesmuschel:Mutator]{miesmuschel::Mutator}} -> \code{\link[miesmuschel:MutatorNumeric]{miesmuschel::MutatorNumeric}} -> \code{MutatorNumericPolynomial}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{MutatorNumericPolynomial$new()}}
\item \href{#method-clone}{\code{MutatorNumericPolynomial$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="miesmuschel" data-topic="MiesOperator" data-id="operate">}\href{../../miesmuschel/html/MiesOperator.html#method-operate}{\code{miesmuschel::MiesOperator$operate()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="miesmuschel" data-topic="MiesOperator" data-id="prime">}\href{../../miesmuschel/html/MiesOperator.html#method-prime}{\code{miesmuschel::MiesOperator$prime()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initialize the \code{MutatorNumericPolynomial} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MutatorNumericPolynomial$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MutatorNumericPolynomial$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
