% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionaries.R
\docType{data}
\name{dict_scalors}
\alias{dict_scalors}
\title{Dictionary of Scalors}
\format{
An object of class \code{DictionaryScalor} (inherits from \code{Dictionary}, \code{R6}) of length 13.
}
\usage{
dict_scalors
}
\description{
Dictionary of Scalors
}
\keyword{datasets}
