% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_mo.R
\name{dist_crowding}
\alias{dist_crowding}
\title{Calculate Crowding Distance}
\usage{
dist_crowding(fitnesses)
}
\arguments{
\item{fitnesses}{(\verb{numeric matrix})\cr
fitness matrix, with one row per individual and one column per objective}
}
\value{
\code{numeric}: Vector of crowding distances.
}
\description{
Takes a \code{matrix} of fitness values and calculates the crowding distance for individuals in that \code{matrix}.

Individuals that are minimal or maximal with respect to at least one dimension are assigned infinite
crowding distance.

Individuals are assumed to be in a (epsilon-) nondominated front.
}
