% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_hb.R
\name{hb_bracket_budget}
\alias{hb_bracket_budget}
\title{Calculate Generalized Hyperband Bracket Budget}
\usage{
hb_bracket_budget(mu, survival_fraction, budgets)
}
\arguments{
\item{mu}{(\code{integer(1)})\cr
Size of first generation}

\item{survival_fraction}{(\code{numeric(1)})\cr
Fraction of generation that gets killed.}

\item{budgets}{(\code{numeric})\cr
budget steps at which individuals get evaluated}
}
\value{
\code{numeric(1)} total SH budget.
}
\description{
Budget needed for evaluation of a single extended successive halving.
}
