% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_hb.R
\name{hb_evaluation_schedule}
\alias{hb_evaluation_schedule}
\title{Create Generalized Hyperband Evaluation Plan}
\usage{
hb_evaluation_schedule(
  mu,
  survival_fraction,
  fidelity_lower,
  fidelity_upper,
  fidelity_steps,
  fidelity_budget_map = exp
)
}
\arguments{
\item{mu}{(\code{integer(1)})\cr
Size of first generation}

\item{survival_fraction}{(\code{numeric(1)})\cr
Fraction of generation that gets killed.}

\item{fidelity_lower}{(\code{numeric(1)})\cr
Lower bound of fidelity parameter}

\item{fidelity_upper}{(\code{numeric(1)})\cr
Upper bound of fidelity parameter}

\item{fidelity_steps}{(\code{integer(1)})\cr
Max number of fidelity steps, at least 1.}

\item{fidelity_budget_map}{(\code{function})\cr
Function that maps fidelity param to budget expenditure. Often this is \code{exp()}}
}
\value{
\code{data.table} with columns \code{fidelity} (\code{numeric}): fidelity component value to evaluateat,
\code{sample_new} (\code{integer}) number of individuals to sample anew, and
\code{survivors} (\code{integer}): number of individuals kept alive from previous stage.
}
\description{
Create a \code{data.table} detailing the number and fidelities of evaluations to perform.
}
