% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_hb.R
\name{hb_survivors_info}
\alias{hb_survivors_info}
\title{Create Generalized Hyperband Survivors Table}
\usage{
hb_survivors_info(mu, survival_fraction, num_stages)
}
\arguments{
\item{mu}{(\code{integer(1)})\cr
Size of first generation}

\item{survival_fraction}{(\code{numeric(1)})\cr
Fraction of generation that gets killed.}

\item{num_stages}{(\code{numeric})\cr
budget steps at which individuals get evaluated}
}
\value{
\code{data.table} with columns \code{survivors} (\code{integer}): number of individuals alive at the start
of the stage, and \code{evaluate} (\code{logical}): whether evaluation is necessary because individuals need
to be killed.
}
\description{
Create a table listing the number of alive individuals at each stage of an SH bracket, as well
as if evaluation is necessary. Evaluation is not necessary if no individual needs to be killed at the
end of a stage (except at the last stage).
}
