% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DensityLearner.R
\name{LearnerDensity}
\alias{LearnerDensity}
\title{Density Estimation Learner}
\description{
This learner implements the \code{\link[mlr3:Learner]{mlr3::Learner}} for density estimation \code{\link[mlr3:Task]{Task}}s.
\itemize{
\item \code{task_type} is set to \code{"density"}.
\item \verb{Creates [}Prediction\verb{][mlr3::Prediction]s of class [}PredictionDensity`].
\item \code{predict_types} always has the value \code{"prob"}.
}
}
\seealso{
Other density estimation classes: 
\code{\link{MeasureDensity}},
\code{\link{PredictionDensity}},
\code{\link{TaskDensity}},
\code{\link{as_prediction_density}()},
\code{\link{as_task_density}()},
\code{\link{mlr_learners_density.featureless}},
\code{\link{mlr_learners_density.np}},
\code{\link{mlr_measures_density.logloss}},
\code{\link{mlr_tasks_faithful}}
}
\concept{density estimation classes}
\section{Super class}{
\code{\link[mlr3:Learner]{mlr3::Learner}} -> \code{LearnerDensity}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{LearnerDensity$new()}}
\item \href{#method-sample}{\code{LearnerDensity$sample()}}
\item \href{#method-clone}{\code{LearnerDensity$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="base_learner">}\href{../../mlr3/html/Learner.html#method-base_learner}{\code{mlr3::Learner$base_learner()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="format">}\href{../../mlr3/html/Learner.html#method-format}{\code{mlr3::Learner$format()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="help">}\href{../../mlr3/html/Learner.html#method-help}{\code{mlr3::Learner$help()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="predict">}\href{../../mlr3/html/Learner.html#method-predict}{\code{mlr3::Learner$predict()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="predict_newdata">}\href{../../mlr3/html/Learner.html#method-predict_newdata}{\code{mlr3::Learner$predict_newdata()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="print">}\href{../../mlr3/html/Learner.html#method-print}{\code{mlr3::Learner$print()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="reset">}\href{../../mlr3/html/Learner.html#method-reset}{\code{mlr3::Learner$reset()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="train">}\href{../../mlr3/html/Learner.html#method-train}{\code{mlr3::Learner$train()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initialize the \code{LearnerDensity} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LearnerDensity$new(
  id,
  param_set = ps(),
  predict_types = "prob",
  feature_types = character(0),
  properties = character(0),
  data_formats = "data.table",
  packages = character(0),
  man = NA_character_
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sample"></a>}}
\if{latex}{\out{\hypertarget{method-sample}{}}}
\subsection{Method \code{sample()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LearnerDensity$sample(n)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LearnerDensity$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
