% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mies_methods.R
\name{mies_evaluate_offspring}
\alias{mies_evaluate_offspring}
\title{Evaluate Proposed Configurations Generated in a MIES Iteration}
\usage{
mies_evaluate_offspring(
  inst,
  offspring,
  fidelity_schedule = NULL,
  budget_id = NULL,
  survivor_budget = FALSE
)
}
\arguments{
\item{inst}{(\code{\link[bbotk:OptimInstance]{OptimInstance}})\cr
Optimization instance to evaluate.}

\item{offspring}{(\code{data.frame})\cr
Proposed configurations to be evaluated, must have columns named after the \code{inst}'s search space, minus \code{budget_id} if not \code{NULL}.}

\item{fidelity_schedule}{(\code{data.frame} | \code{NULL})\cr
\code{data.frame} with three columns \code{"generation"}, \code{"budget_new"}, \code{"budget_survivors"}, in that order. \code{"budget_new"} and \code{"budget_survivors"}
are atomic columns assigned to the \code{budget_id} component of \code{offspring}; which one is chosen depends on \code{survivor_budget}.
\code{"generation"} is an integer valued column, indicating the first generation at which a row is valid. At least one row with
\code{generation == 1} must be present.\cr
This value may be \code{NULL} if no multi-fidelity optimization is performed (the default).}

\item{budget_id}{(\code{character(1)} | \code{NULL})\cr
Budget component that is set to the value found in \code{fidelity_schedule} when doing multi-fidelity optimization. Should be \code{NULL}
when and only when \code{fidelity_schedule} is \code{NULL}.}

\item{survivor_budget}{(\code{logical(1)})\cr
When doing multi-fidelity optimization, determines which column of \code{fidelity_schedule} to use to determine the budget component value.}
}
\value{
invible \code{\link[data.table:data.table]{data.table}}: the performance values returned when evaluating the \code{offspring} values
through \code{eval_batch}.
}
\description{
Calls \verb{$eval_batch} of a given \code{\link[bbotk:OptimInstance]{OptimInstance}} on a set
of configurations as part of a MIES operation. The \code{dob} extra-info in the archive
is also set properly to indicate a progressed generation.

This function can be used directly, but it is easier to use it within the \code{\link{OptimizerMies}}
\code{\link[bbotk:Optimizer]{Optimizer}} if standard GA operation is desired.

"Rolling-tide" multifidelity is supported as described in \code{\link{OptimizerMies}} or more thoroughly
in \code{vignette("mies-multifid")}. For this,
an extra component named after \code{budget_id} is appended to each individual, chosen from
the \code{fidelity_schedule} depending on the value of \code{survivor_budget}. Both \code{fidelity_schedule}
and \code{budget_id} should be the same values as given to the other \verb{mies_*} functions.
}
\seealso{
Other mies building blocks: 
\code{\link{mies_generate_offspring}()},
\code{\link{mies_get_fitnesses}()},
\code{\link{mies_init_population}()},
\code{\link{mies_select_from_archive}()},
\code{\link{mies_step_fidelity}()},
\code{\link{mies_survival_comma}()},
\code{\link{mies_survival_plus}()}
}
\concept{mies building blocks}
