% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mies_methods.R
\name{mies_get_fitnesses}
\alias{mies_get_fitnesses}
\title{Get Fitness Values from OptimInstance}
\usage{
mies_get_fitnesses(inst, rows)
}
\arguments{
\item{inst}{(\code{\link[bbotk:OptimInstance]{OptimInstance}})\cr
Optimization instance to evaluate.}

\item{rows}{optional (\code{integer})\cr
Indices of rows within \code{inst} to consider. If this is not given, then the entire archive is used.}
}
\value{
\code{numeric} \code{matrix} with \code{length(rows)} (if \code{rows} is given, otherwise \code{nrow(inst$archive$data)}) rows
and one column for each objective: fitnesses to be maximized.
}
\description{
Get fitness values in the correct form as used by \code{\link{Selector}} operators from an
\code{\link[bbotk:OptimInstance]{OptimInstance}}.
This works for both single-criterion and multi-criterion optimization, and entails multiplying
objectives with -1 if they are being minimized, since \code{\link{Selector}} tries to maximize fitness.
}
\seealso{
Other mies building blocks: 
\code{\link{mies_evaluate_offspring}()},
\code{\link{mies_generate_offspring}()},
\code{\link{mies_init_population}()},
\code{\link{mies_select_from_archive}()},
\code{\link{mies_step_fidelity}()},
\code{\link{mies_survival_comma}()},
\code{\link{mies_survival_plus}()}
}
\concept{mies building blocks}
