% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mies_methods.R
\name{mies_select_from_archive}
\alias{mies_select_from_archive}
\title{Select Individuals from an OptimInstance}
\usage{
mies_select_from_archive(
  inst,
  n_select,
  rows,
  selector = SelectorBest$new()$prime(inst$search_space),
  get_indivs = TRUE
)
}
\arguments{
\item{inst}{(\code{\link[bbotk:OptimInstance]{OptimInstance}})\cr
Optimization instance to evaluate.}

\item{n_select}{(\code{integer(1)})\cr
Number of individuals to select.}

\item{rows}{optional (\code{integer})\cr
Indices of rows within \code{inst} to consider. If this is not given, then the entire archive is used.}

\item{get_indivs}{(\code{logical(1)})\cr
Whether to return configuration values from within the archive (\code{TRUE}) or just the indices within
the archive (\code{FALSE}). Default is \code{TRUE}.}

\item{`selector`}{(\code{\link{Selector}})\cr
\code{\link{Selector}} operator that selects individuals depending on configuration values
and objective results. When \code{selector$operate()} is called, then objectives that
are being minimized are multiplied with -1 (through \code{\link[=mies_get_fitnesses]{mies_get_fitnesses()}}), since \code{\link{Selector}}s always try to maximize fitness.
Defaults to \code{\link{SelectorBest}}.\cr
The \code{\link{Selector}} must be primed on a superset of \code{inst$search_space}; this \emph{includes} the "budget" component
when performing multi-fidelity optimization. All components on which \code{selector} is primed on must occur in the archive.\cr
The given \code{\link{Selector}} \emph{may} return duplicates.}
}
\value{
\code{integer} | \code{\link[data.table:data.table]{data.table}}: Selected individuals, either index into \code{inst} or subset of archive table,
depending on \code{get_indivs}.
}
\description{
Apply a \code{\link{Selector}} operator to a subset of configurations inside
an \code{\link[bbotk:OptimInstance]{OptimInstance}}
and return the index within the archive (when \code{get_indivs} \code{FALSE}) or the configurations themselves
(when \code{get_indivs} is \code{TRUE}).

It is not strictly necessary for the selector to select unique individuals / individuals without replacement.
}
\seealso{
Other mies building blocks: 
\code{\link{mies_evaluate_offspring}()},
\code{\link{mies_generate_offspring}()},
\code{\link{mies_get_fitnesses}()},
\code{\link{mies_init_population}()},
\code{\link{mies_step_fidelity}()},
\code{\link{mies_survival_comma}()},
\code{\link{mies_survival_plus}()}
}
\concept{mies building blocks}
