% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{Miesmuschel-package}
\alias{Miesmuschel}
\alias{Miesmuschel-package}
\title{Miesmuschel: Mixed Integer Evolutionary Strategies.}
\description{

}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mlr-org/miesmuschel}
  \item Report bugs at \url{https://github.com/mlr-org/miesmuschel/issues}
}

}
\author{
\strong{Maintainer}: Martin Binder \email{mlr.developer@mb706.com}

}
