% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LearnerDensityNP.R
\name{mlr_learners_density.np}
\alias{mlr_learners_density.np}
\alias{LearnerDensityNP}
\title{Kernel Density Estimator using the "np" Package}
\description{
Perform kernel density estimation using the \code{np} package.
}
\section{Hyperparameters}{

\itemize{
\item \code{bwmethod} :: \code{character(1)}\cr
Bandwidth selection method. One of \code{"cv.ml"}, \code{"cv.ls"}, \code{"normal-reference"}, \code{"normal-reference-numeric"}. Default \code{"cv.ml"}.
\code{"normal-reference-numeric"} emulates (buggy, see \verb{https://github.com/statsmodels/statsmodels/issues/3790}) behaviour of python statsmodels:
non-numeric columns are treated as numeric for bandwidth estimation.
\item \code{bwtype} :: \code{character(1)}\cr
Continuous variable bandwidth type, one of \code{"fixed"} (default), \code{"generalized_nn"}, \code{"adaptive_nn"}.
\item \code{ckertype} :: \code{character(1)}\cr
Continuous kernel type. One of \code{"gaussian"} (default), \code{"epanechnikov"}, \code{"uniform"}.
\item \code{ckerorder} :: \code{integer(1)}\cr
Continuous Kernel order (only for gaussian and epanechnikov). One of 2 (default), 4, 6, 8.
\item \code{ukertype} :: \code{character(1)}\cr
Unordered categorical kernel type. One of \code{"aitchisonaitken"} (default) or \code{"liracine"}.
\item \code{okertype} :: \code{character(1)}\cr
Ordered kernel type. One of \code{"liracine"} (default), \code{"wangvanryzin"}.
\item \code{nmulti} :: \code{integer(1)}\cr
Number of random restarts for cross validation likelihood optimization. Default: number of features, at most 5. 0 disables restarts.
\item \code{remin} :: \code{logical(1)}\cr
Restart from located minima; Default \code{TRUE}.
\item \code{itmax} :: \code{integer(1)}\cr
Max optimization iters. Default 10000.
\item \code{ftol} :: \code{numeric(1)}\cr
Optimization y-value relative tolerance, default \code{10 * sqrt(.Machine$double.eps)}.
\item \code{tol} :: \code{numeric(1)}\cr
Optimization x-value relative tolerance, default \code{10000 * sqrt(.Machine$double.eps)}.
\item \code{small} :: \code{numeric(1)}\cr
Optimization x-value absolute tolerance (?), default \code{1000 * sqrt(.Machine$double.eps)}.
\item \code{min_bandwidth} :: \code{numeric(1)}\cr
Minimum bandwidth (all kerneltypes). Not part of np::npudensbw. Default 0. BOHB has this at \code{1e-3}.
\item \code{sampling_bw_factor} :: \code{numeric(1)}\cr
Oversmoothing bandwidth factor for sampling. Not part of np::npudensbw. Default 1. BOHB has this at 3.
}

TODO: could also implement \verb{lb\{c,d\}.\{dir,init\}}, \verb{\{c,d\}fac.\{dir,init\}}, \code{dfc.dir}, \code{hbd.dir}, \verb{hb\{c,d\}.init}, \verb{init\{c,d\}.dir}, \code{scale.init.categorical.sample}.

TODO: \code{normal-reference} seems to disable optimization, consider putting that as a dependency on the optimization-related params.

TODO: could add dependency for ckerorder on not-uniform.
}

\seealso{
Other density estimation classes: 
\code{\link{LearnerDensity}},
\code{\link{MeasureDensity}},
\code{\link{PredictionDensity}},
\code{\link{TaskDensity}},
\code{\link{as_prediction_density}()},
\code{\link{as_task_density}()},
\code{\link{mlr_learners_density.featureless}},
\code{\link{mlr_measures_density.logloss}},
\code{\link{mlr_tasks_faithful}}
}
\concept{density estimation classes}
\section{Super classes}{
\code{\link[mlr3:Learner]{mlr3::Learner}} -> \code{\link[miesmuschel:LearnerDensity]{miesmuschel::LearnerDensity}} -> \code{LearnerDensityNP}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{LearnerDensityNP$new()}}
\item \href{#method-clone}{\code{LearnerDensityNP$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="base_learner">}\href{../../mlr3/html/Learner.html#method-base_learner}{\code{mlr3::Learner$base_learner()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="format">}\href{../../mlr3/html/Learner.html#method-format}{\code{mlr3::Learner$format()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="help">}\href{../../mlr3/html/Learner.html#method-help}{\code{mlr3::Learner$help()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="predict">}\href{../../mlr3/html/Learner.html#method-predict}{\code{mlr3::Learner$predict()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="predict_newdata">}\href{../../mlr3/html/Learner.html#method-predict_newdata}{\code{mlr3::Learner$predict_newdata()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="print">}\href{../../mlr3/html/Learner.html#method-print}{\code{mlr3::Learner$print()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="reset">}\href{../../mlr3/html/Learner.html#method-reset}{\code{mlr3::Learner$reset()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="train">}\href{../../mlr3/html/Learner.html#method-train}{\code{mlr3::Learner$train()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="miesmuschel" data-topic="LearnerDensity" data-id="sample">}\href{../../miesmuschel/html/LearnerDensity.html#method-sample}{\code{miesmuschel::LearnerDensity$sample()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initialize the \code{LearnerDensityNP} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LearnerDensityNP$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LearnerDensityNP$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
