% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PipeOpDensityRatio.R
\name{mlr_pipeops_densityratio}
\alias{mlr_pipeops_densityratio}
\alias{PipeOpDensityRatio}
\title{Predict Ratio of Densities}
\format{
\code{\link{R6Class}} inheriting from \code{\link{PipeOp}}.
}
\description{
Predict the ratio of two \code{\link{PredictionDensity}} inputs as \code{\link{PredictionRegr}}, dividing the input to \code{"numerator"} by the input to \code{"denominator"}.
}
\section{Construction}{
\preformatted{PipeOpDensityRatio$new(id = "densityratio", param_vals = list())
}
\itemize{
\item \code{id} :: \code{character(1)}\cr
Identifier of the resulting object, default \code{"densityratio"}.
\item \code{param_vals} :: named \code{list}\cr
List of hyperparameter settings, overwriting the hyperparameter settings that would otherwise be set during construction. Default \code{list()}.
}
}

\section{Input and Output Channels}{

\code{\link{PipeOpDensityRatio}} has two input channels named \code{"numerator"} and \code{"denominator"}, both taking \code{NULL} during training and a \code{\link{PredictionDensity}}
during prediction.

\code{\link{PipeOpDensityRatio}} has one output channel named \code{"output"}, returning \code{NULL} during training and a \code{\link[mlr3:PredictionRegr]{PredictionRegr}}
during prediction.
}

\section{State}{

The \verb{$state} is left empty (\code{list()}).
}

\section{Parameters}{

The parameters are the parameters inherited from the \code{\link{PipeOp}}, as well as:
\itemize{
\item \code{min_density} :: \code{numeric(1)}\cr
Minimum density to consider for density ratio. Should be marginally greater than 0 to avoid too large numbers when dividing. Initialized to \code{1e-32}.
}
}

\section{Internals}{

Can best be used with \code{\link{PipeOpDensitySplit}}
}

\section{Fields}{

Only fields inherited from \code{\link{PipeOp}}.
}

\section{Methods}{

Only methods inherited from \code{\link{PipeOp}}.
}

\seealso{
Other PipeOps: 
\code{\link{mlr_pipeops_densitysplit}},
\code{\link{mlr_pipeops_predictionunion}},
\code{\link{mlr_pipeops_stratify}}

Other BOHB implementing operations: 
\code{\link{mlr_pipeops_densitysplit}}
}
\concept{BOHB implementing operations}
\concept{Ensembles}
\concept{PipeOps}
\section{Super class}{
\code{\link[mlr3pipelines:PipeOp]{mlr3pipelines::PipeOp}} -> \code{PipeOpDensityRatio}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{PipeOpDensityRatio$new()}}
\item \href{#method-clone}{\code{PipeOpDensityRatio$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="mlr3pipelines" data-topic="PipeOp" data-id="predict">}\href{../../mlr3pipelines/html/PipeOp.html#method-predict}{\code{mlr3pipelines::PipeOp$predict()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3pipelines" data-topic="PipeOp" data-id="print">}\href{../../mlr3pipelines/html/PipeOp.html#method-print}{\code{mlr3pipelines::PipeOp$print()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3pipelines" data-topic="PipeOp" data-id="train">}\href{../../mlr3pipelines/html/PipeOp.html#method-train}{\code{mlr3pipelines::PipeOp$train()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PipeOpDensityRatio$new(id = "densityratio", param_vals = list())}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PipeOpDensityRatio$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
