% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionaries.R
\docType{data}
\name{mlr_recombinators}
\alias{mlr_recombinators}
\title{Dictionary of Recombinators}
\format{
An object of class \code{DictionaryRecombinator} (inherits from \code{Dictionary}, \code{R6}) of length 13.
}
\usage{
mlr_recombinators
}
\description{
Dictionary of Recombinators
}
\keyword{datasets}
