% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionaries.R
\docType{data}
\name{mlr_selectors}
\alias{mlr_selectors}
\title{Dictionary of Selectors}
\format{
An object of class \code{DictionarySelector} (inherits from \code{Dictionary}, \code{R6}) of length 13.
}
\usage{
mlr_selectors
}
\description{
Dictionary of Selectors
}
\keyword{datasets}
