% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DensityTask.R
\name{mlr_tasks_faithful}
\alias{mlr_tasks_faithful}
\title{Old Faithful Density Estimation Task}
\format{
\link[R6:R6Class]{R6::R6Class} inheriting from \code{\link{TaskDens}}
}
\description{
A density estimation task for the \code{\link[datasets:faithful]{datasets::faithful}} dataset.
}
\section{Construction}{
\preformatted{tsk("faithful")
}
}

\seealso{
Other density estimation classes: 
\code{\link{LearnerDensity}},
\code{\link{MeasureDensity}},
\code{\link{PredictionDensity}},
\code{\link{TaskDensity}},
\code{\link{as_prediction_density}()},
\code{\link{as_task_density}()},
\code{\link{mlr_learners_density.featureless}},
\code{\link{mlr_learners_density.np}},
\code{\link{mlr_measures_density.logloss}}
}
\concept{density estimation classes}
