% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DensityPrediction.R
\name{PredictionDensity}
\alias{PredictionDensity}
\title{Density Estimation Prediction Object}
\description{
Predictions returned by \code{\link{LearnerDensity}} and evaluated by \code{\link{MeasureDensity}}.
}
\seealso{
Other density estimation classes: 
\code{\link{LearnerDensity}},
\code{\link{MeasureDensity}},
\code{\link{TaskDensity}},
\code{\link{as_prediction_density}()},
\code{\link{as_task_density}()},
\code{\link{mlr_learners_density.featureless}},
\code{\link{mlr_learners_density.np}},
\code{\link{mlr_measures_density.logloss}},
\code{\link{mlr_tasks_faithful}}
}
\concept{density estimation classes}
\section{Super class}{
\code{\link[mlr3:Prediction]{mlr3::Prediction}} -> \code{PredictionDensity}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{prob}}{(\code{numeric})\cr
Predicted density.}

\item{\code{prob}}{(\code{numeric})\cr
Predicted log density.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{PredictionDensity$new()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="mlr3" data-topic="Prediction" data-id="filter">}\href{../../mlr3/html/Prediction.html#method-filter}{\code{mlr3::Prediction$filter()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3" data-topic="Prediction" data-id="format">}\href{../../mlr3/html/Prediction.html#method-format}{\code{mlr3::Prediction$format()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3" data-topic="Prediction" data-id="help">}\href{../../mlr3/html/Prediction.html#method-help}{\code{mlr3::Prediction$help()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3" data-topic="Prediction" data-id="print">}\href{../../mlr3/html/Prediction.html#method-print}{\code{mlr3::Prediction$print()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3" data-topic="Prediction" data-id="score">}\href{../../mlr3/html/Prediction.html#method-score}{\code{mlr3::Prediction$score()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initialize the \code{PredictionDensity} object. One of \code{prob} or \code{logprob} must be given.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PredictionDensity$new(
  task = NULL,
  row_ids = mlr3::assert_task(task)$row_ids,
  prob = exp(logprob),
  logprob = log(prob),
  check = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{task}}{(\code{\link{TaskDensity}})\cr
\code{\link{TaskDensity}} for which prediction was made, used to extract defaults for \code{row_ids}.}

\item{\code{row_ids}}{(\code{integer})\cr
Row ids of the predicted observations.}

\item{\code{prob}}{(\code{numeric})\cr
Predicted density}

\item{\code{logprob}}{(\code{numeric})\cr
Predicted log density.}

\item{\code{check}}{(\code{logical(1)})\cr
Whether to perform argument checks and predict type conversions.}
}
\if{html}{\out{</div>}}
}
}
}
