% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TerminatorGenerations.R
\name{terminator_get_generations}
\alias{terminator_get_generations}
\title{Get the Numger of Generations that a Terminator Allows}
\usage{
terminator_get_generations(x)
}
\description{
Get the number of generations of a \code{\link{TerminatorGenerations}}. When the \code{\link{TerminatorGenerations}}
is wrapped in a \code{\link[bbotk:mlr_terminators_combo]{TerminatorCombo}}, then the minimum number of generations
allowed by it are retrieved. This is the minimum of all \code{terminator_get_generations} if \verb{$any} is set
to \code{TRUE}, and the maximum if \verb{$any} is set to \code{FALSE}.

The number of generations allowed by other \code{\link[bbotk:Terminator]{Terminator}}s is infinity.
}
