% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{TuningInstanceMultiCrit}
\alias{TuningInstanceMultiCrit}
\title{TuningInstanceMultiCrit Class}
\description{
\code{mlr3tuning}'s \code{TuningInstanceMultiCrit} class.
Re-exported since \code{mlr3tuning} will change the name.
}
\section{Super classes}{
\code{\link[bbotk:OptimInstance]{bbotk::OptimInstance}} -> \code{\link[bbotk:OptimInstanceMultiCrit]{bbotk::OptimInstanceMultiCrit}} -> \code{\link[mlr3tuning:TuningInstanceMultiCrit]{mlr3tuning::TuningInstanceMultiCrit}} -> \code{TuningInstanceMultiCrit}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TuningInstanceMultiCrit-clone}{\code{TuningInstanceMultiCrit$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="bbotk" data-topic="OptimInstance" data-id="clear"><a href='../../bbotk/html/OptimInstance.html#method-OptimInstance-clear'><code>bbotk::OptimInstance$clear()</code></a></span></li>
<li><span class="pkg-link" data-pkg="bbotk" data-topic="OptimInstance" data-id="eval_batch"><a href='../../bbotk/html/OptimInstance.html#method-OptimInstance-eval_batch'><code>bbotk::OptimInstance$eval_batch()</code></a></span></li>
<li><span class="pkg-link" data-pkg="bbotk" data-topic="OptimInstance" data-id="format"><a href='../../bbotk/html/OptimInstance.html#method-OptimInstance-format'><code>bbotk::OptimInstance$format()</code></a></span></li>
<li><span class="pkg-link" data-pkg="bbotk" data-topic="OptimInstance" data-id="objective_function"><a href='../../bbotk/html/OptimInstance.html#method-OptimInstance-objective_function'><code>bbotk::OptimInstance$objective_function()</code></a></span></li>
<li><span class="pkg-link" data-pkg="bbotk" data-topic="OptimInstance" data-id="print"><a href='../../bbotk/html/OptimInstance.html#method-OptimInstance-print'><code>bbotk::OptimInstance$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3tuning" data-topic="TuningInstanceMultiCrit" data-id="assign_result"><a href='../../mlr3tuning/html/TuningInstanceMultiCrit.html#method-TuningInstanceMultiCrit-assign_result'><code>mlr3tuning::TuningInstanceMultiCrit$assign_result()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3tuning" data-topic="TuningInstanceMultiCrit" data-id="initialize"><a href='../../mlr3tuning/html/TuningInstanceMultiCrit.html#method-TuningInstanceMultiCrit-initialize'><code>mlr3tuning::TuningInstanceMultiCrit$initialize()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TuningInstanceMultiCrit-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TuningInstanceMultiCrit-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TuningInstanceMultiCrit$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
