% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mies_methods.R
\name{mies_filter_offspring}
\alias{mies_filter_offspring}
\title{Filter Offspring}
\usage{
mies_filter_offspring(
  inst,
  individuals,
  lambda,
  filtor = NULL,
  get_indivs = TRUE
)
}
\arguments{
\item{inst}{(\code{\link[bbotk:OptimInstance]{OptimInstance}})\cr
Optimization instance to evaluate.}

\item{lambda}{(\code{integer(1)})\cr
Number of individuals to filter down to.}

\item{filtor}{(\code{\link{Filtor}} | \code{NULL})\cr
\code{\link{Filtor}} operator that filters.}

\item{get_indivs}{(\code{logical(1)})\cr
Whether to return the selected individuals, or an index into \code{individuals}.}
}
\value{
If \code{get_indivs} is \code{TRUE}: a \code{data.frame} or \code{\link[data.table:data.table]{data.table}} (depending on the type of \code{individuals}) of filtered configurations.
Otherwise: an integer vector indexing the filtered individuals.
}
\description{
TODO
}
