% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiesOperator.R
\name{MiesOperator}
\alias{MiesOperator}
\title{MiesOperator}
\description{
MiesOperator

MiesOperator
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{MiesOperator$new()}}
\item \href{#method-prime}{\code{MiesOperator$prime()}}
\item \href{#method-operate}{\code{MiesOperator$operate()}}
\item \href{#method-clone}{\code{MiesOperator$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MiesOperator$new(
  param_classes = c("ParamLgl", "ParamInt", "ParamDbl", "ParamFct"),
  param_set = ps()
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-prime"></a>}}
\if{latex}{\out{\hypertarget{method-prime}{}}}
\subsection{Method \code{prime()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MiesOperator$prime(param_set)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-operate"></a>}}
\if{latex}{\out{\hypertarget{method-operate}{}}}
\subsection{Method \code{operate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MiesOperator$operate(values, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MiesOperator$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
