% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionaries.R
\name{mut}
\alias{mut}
\alias{muts}
\alias{rec}
\alias{recs}
\alias{sel}
\alias{sels}
\title{Short Access Forms for Operators}
\usage{
mut(.key, ...)

muts(.keys, ...)

rec(.key, ...)

recs(.key, ...)

sel(.key, ...)

sels(.key, ...)
}
\arguments{
\item{.key}{(\code{character(1)})\cr
Key passed to the respective \link[mlr3misc:Dictionary]{dictionary} to retrieve the object.}

\item{...}{(named \code{list()})\cr
Named arguments passed to the constructor, to be set as parameters in the \link[paradox:ParamSet]{paradox::ParamSet}, or to be set as public field.
See \code{\link[mlr3misc:dictionary_sugar_get]{mlr3misc::dictionary_sugar_get()}} for more details.}

\item{.keys}{(\code{character()})\cr
Keys passed to the respective \link[mlr3misc:Dictionary]{dictionary} to retrieve multiple objects.}
}
\value{
\itemize{
\item \code{\link{Mutator}} for \code{mut()}
\item list of \code{\link{Mutator}} for \code{muts()}
\item \code{\link{Recombinator}} for \code{rec()}.
\item list of \code{\link{Recombinator}} for \code{recs()}.
\item \code{\link{Selector}} for \code{sel()}.
\item list of \code{\link{Selector}} for \code{sels()}.
}
}
\description{
These functions complement \link{dict_mutators}, \link{dict_recombinators}, \link{dict_selectors} with functions in the spirit
of \link[mlr3:mlr_sugar]{mlr3::mlr_sugar}.
}
\examples{
mut("gauss", sdev = 0.5)
rec("xounif")
sel("random")
}
