% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{OptimInstanceMultiCrit}
\alias{OptimInstanceMultiCrit}
\title{OptimInstanceMultiCrit Class}
\description{
\code{bbotk}'s \code{OptimInstanceMultiCrit} class.
Re-exported since \code{bbotk} will change the name.
}
\section{Super classes}{
\code{bbotk::OptimInstance} -> \code{bbotk::OptimInstanceBatch} -> \code{bbotk::OptimInstanceBatchMultiCrit} -> \code{OptimInstanceMultiCrit}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OptimInstanceMultiCrit-clone}{\code{OptimInstanceMultiCrit$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="bbotk" data-topic="OptimInstance" data-id="clear"><a href='../../bbotk/html/OptimInstance.html#method-OptimInstance-clear'><code>bbotk::OptimInstance$clear()</code></a></span></li>
<li><span class="pkg-link" data-pkg="bbotk" data-topic="OptimInstance" data-id="format"><a href='../../bbotk/html/OptimInstance.html#method-OptimInstance-format'><code>bbotk::OptimInstance$format()</code></a></span></li>
<li><span class="pkg-link" data-pkg="bbotk" data-topic="OptimInstance" data-id="print"><a href='../../bbotk/html/OptimInstance.html#method-OptimInstance-print'><code>bbotk::OptimInstance$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="bbotk" data-topic="OptimInstanceBatch" data-id="eval_batch"><a href='../../bbotk/html/OptimInstanceBatch.html#method-OptimInstanceBatch-eval_batch'><code>bbotk::OptimInstanceBatch$eval_batch()</code></a></span></li>
<li><span class="pkg-link" data-pkg="bbotk" data-topic="OptimInstanceBatch" data-id="objective_function"><a href='../../bbotk/html/OptimInstanceBatch.html#method-OptimInstanceBatch-objective_function'><code>bbotk::OptimInstanceBatch$objective_function()</code></a></span></li>
<li><span class="pkg-link" data-pkg="bbotk" data-topic="OptimInstanceBatchMultiCrit" data-id="assign_result"><a href='../../bbotk/html/OptimInstanceBatchMultiCrit.html#method-OptimInstanceBatchMultiCrit-assign_result'><code>bbotk::OptimInstanceBatchMultiCrit$assign_result()</code></a></span></li>
<li><span class="pkg-link" data-pkg="bbotk" data-topic="OptimInstanceBatchMultiCrit" data-id="initialize"><a href='../../bbotk/html/OptimInstanceBatchMultiCrit.html#method-OptimInstanceBatchMultiCrit-initialize'><code>bbotk::OptimInstanceBatchMultiCrit$initialize()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OptimInstanceMultiCrit-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OptimInstanceMultiCrit-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OptimInstanceMultiCrit$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
