% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OptimizerMies.R
\name{OptimizerMies}
\alias{OptimizerMies}
\title{Mixed Integer Evolutionary Strategies Optimizer}
\description{
Perform optimization using evolutionary strategies. \code{OptimizerMies} and \code{TunerMies} implement a standard ES optimization
algorithm, performing initialization first, followed by a loop of performance evaluation, survival selection, parent selection, mutation, and
recombination. Currently two different survival modes ("comma" and "plus") are supported. Multi-fidelity optimization, similar
to the "rolling-tide" algorithm described in Fieldsend (2014). The modular design and reliance on
\code{\link{MiesOperator}} objects to perform central parts of the optimization algorithm makes this \code{\link[bbotk:Optimizer]{Optimizer}}
highly flexible and configurable. In combination with \code{\link{OperatorCombination}} mutators and recombinators, an algorithm
as presented in Li (2013) can easily be implemented.

\code{OptimizerMies} implements a standard evolutionary strategies loop:
\enumerate{
\item Prime operators, using \code{mies_prime_operators()}
\item Initialize population, using \code{mies_init_population()}
\item Generate offspring by selecting parents, recombining and mutating them, using \code{mies_generate_offspring()}
\item Evaluate performance, using \code{mies_evaluate_offspring()}
\item Select survivors, using either \code{mies_survival_plus()} or \code{mies_survival_comma()}, depending on the \code{survival_strategy} hyperparameter
\item Optionally, evaluate survivors with higher fidelity if the multi-fidelity functionality is being used
\item Jump to 3.
}
}
\section{Multi-Fidelity}{

\code{miesmuschel} provides a simple multi-fidelity optimization mechanism that both allows increasing fidelity by generation
number, and
}

\references{
Fieldsend, E J, Everson, M R (2014).
\dQuote{The rolling tide evolutionary algorithm: A multiobjective optimizer for noisy optimization problems.}
\emph{IEEE Transactions on Evolutionary Computation}, \bold{19}(1), 103--117.

Li, Rui, Emmerich, TM M, Eggermont, Jeroen, B"ack, Thomas, Sch"utz, Martin, Dijkstra, Jouke, Reiber, HC J (2013).
\dQuote{Mixed integer evolution strategies for parameter optimization.}
\emph{Evolutionary computation}, \bold{21}(1), 29--64.
}
\section{Super class}{
\code{\link[bbotk:Optimizer]{bbotk::Optimizer}} -> \code{OptimizerMies}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{OptimizerMies$new()}}
\item \href{#method-clone}{\code{OptimizerMies$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="bbotk" data-topic="Optimizer" data-id="format">}\href{../../bbotk/html/Optimizer.html#method-format}{\code{bbotk::Optimizer$format()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="bbotk" data-topic="Optimizer" data-id="optimize">}\href{../../bbotk/html/Optimizer.html#method-optimize}{\code{bbotk::Optimizer$optimize()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="bbotk" data-topic="Optimizer" data-id="print">}\href{../../bbotk/html/Optimizer.html#method-print}{\code{bbotk::Optimizer$print()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OptimizerMies$new(
  mutator = MutatorProxy$new(),
  recombinator = RecombinatorProxy$new(),
  parent_selector = SelectorProxy$new(),
  survival_selector = SelectorProxy$new(),
  elite_selector = NULL,
  multi_fidelity = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OptimizerMies$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
