% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RecombinatorProxy.R
\name{RecombinatorProxy}
\alias{RecombinatorProxy}
\title{Proxy-Recombinator that Recombines According to its Configuration parameter}
\description{
Recombinator that performs the operation in its \code{operation} configuration parameter. This is useful, e.g., to make
\code{\link{OptimizerMies}}'s recombination operation fully parametrizable.

Changes in the \code{operation} configuration parameter are only realized whenever \verb{$prime()} is called, so \verb{$prime()}
must be called every time when \code{operation} is changed, \emph{even if} the new configuration parameter value is already primed.
}
\section{Configuration Parameters}{

\itemize{
\item \code{operation} :: \code{\link{Recombinator}}\cr
Operation to perform. Initialized to \code{\link{RecombinatorNull}} with appropriate \code{n_indivs_in} and \code{n_indivs_out} values.
}
}

\section{Supported Operand Types}{


Supported \code{\link[paradox:Param]{Param}} classes are: \code{\link[paradox:ParamLgl]{ParamLgl}}, \code{\link[paradox:ParamInt]{ParamInt}}, \code{\link[paradox:ParamDbl]{ParamDbl}}, \code{\link[paradox:ParamFct]{ParamFct}}
}

\section{Dictionary}{

This \code{\link{Recombinator}} can be created with the short access form \code{\link[=rec]{rec()}}
(\code{\link[=recs]{recs()}} to get a list), or through the the \link[mlr3misc:Dictionary]{dictionary}
\code{\link{dict_recombinators}} in the following way:\preformatted{# preferred:
rec("proxy")
recs("proxy")  # takes vector IDs, returns list of Recombinators

# long form:
dict_recombinators$get("proxy")
}
}

\seealso{
Other recombinators: 
\code{\link{OperatorCombination}},
\code{\link{Recombinator}},
\code{\link{dict_recombinators_maybe}},
\code{\link{dict_recombinators_null}},
\code{\link{dict_recombinators_xounif}}

Other recombinator wrappers: 
\code{\link{OperatorCombination}},
\code{\link{dict_recombinators_maybe}}
}
\concept{recombinator wrappers}
\concept{recombinators}
\section{Super classes}{
\code{\link[miesmuschel:MiesOperator]{miesmuschel::MiesOperator}} -> \code{\link[miesmuschel:Recombinator]{miesmuschel::Recombinator}} -> \code{RecombinatorProxy}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{RecombinatorProxy$new()}}
\item \href{#method-prime}{\code{RecombinatorProxy$prime()}}
\item \href{#method-clone}{\code{RecombinatorProxy$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="miesmuschel" data-topic="MiesOperator" data-id="operate">}\href{../../miesmuschel/html/MiesOperator.html#method-operate}{\code{miesmuschel::MiesOperator$operate()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initialize the \code{RecombinatorProxy} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RecombinatorProxy$new(n_indivs_in = 2, n_indivs_out = n_indivs_in)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n_indivs_in}}{(\code{integer(1)})\cr
Number of individuals to consider at the same time. When operating, the number of input individuals must be divisible by this number.
Furthermore, the \code{\link{Recombinator}} assigned to the \code{operation} configuration parameter must have an \code{n_indivs_in} that is a divisor of this number.
Default 2.\cr
The \verb{$n_indivs_in} field will reflect this value.}

\item{\code{n_indivs_out}}{(\code{integer(1)})\cr
Number of individuals that result for each \code{n_indivs_in} lines of input. Must be at most \code{n_indivs_in}.
The ratio of \verb{$n_indivs_in} to \verb{$n_indivs_out} of the \code{\link{Recombinator}} assigned to the \code{operation} configuration parameter must be the same as
\code{n_indivs_in} to \code{n_indivs_out} of this object.
Default equal to \code{n_indivs_in}.\cr
The \verb{$n_indivs_out} field will reflect this value.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-prime"></a>}}
\if{latex}{\out{\hypertarget{method-prime}{}}}
\subsection{Method \code{prime()}}{
See \code{\link{MiesOperator}} method. Primes both this operator, as well as the operator given to the \code{operation} configuration parameter.
This must be called whenever the \code{operation} configuration parameter changes, \emph{even if} the configuration parameter is already primed.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RecombinatorProxy$prime(param_set)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{param_set}}{(\code{\link[paradox:ParamSet]{ParamSet}})\cr
Passed to \code{\link{MiesOperator}}\verb{$prime()}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\link{invisible} \code{self}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RecombinatorProxy$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
