% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TerminatorGenerations.R
\name{terminator_get_generations}
\alias{terminator_get_generations}
\title{Get the Numger of Generations that a Terminator Allows}
\usage{
terminator_get_generations(x)
}
\arguments{
\item{x}{(\code{Terminator})\cr
\code{Terminator} to query.}
}
\value{
\code{numeric(1)}: The theoretical maximum number of generations allowed by the \code{Terminator}.
}
\description{
Get the number of generations of a \code{\link{TerminatorGenerations}}. When the \code{\link{TerminatorGenerations}}
is wrapped in a \code{TerminatorCombo}, then the minimum number of generations
allowed by it are retrieved. This is the minimum of all \code{terminator_get_generations} if \verb{$any} is set
to \code{TRUE}, and the maximum if \verb{$any} is set to \code{FALSE}.

The number of generations allowed by other \code{Terminator}s is infinity.
}
