% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{TuningInstanceSingleCrit}
\alias{TuningInstanceSingleCrit}
\title{TuningInstanceSingleCrit Class}
\description{
\code{mlr3tuning}'s \code{TuningInstanceSingleCrit} class.
Re-exported since \code{mlr3tuning} will change the name.
}
\section{Super classes}{
\code{\link[bbotk:OptimInstance]{bbotk::OptimInstance}} -> \code{\link[bbotk:OptimInstanceBatch]{bbotk::OptimInstanceBatch}} -> \code{\link[bbotk:OptimInstanceBatchSingleCrit]{bbotk::OptimInstanceBatchSingleCrit}} -> \code{\link[mlr3tuning:TuningInstanceBatchSingleCrit]{mlr3tuning::TuningInstanceBatchSingleCrit}} -> \code{TuningInstanceSingleCrit}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TuningInstanceSingleCrit-clone}{\code{TuningInstanceSingleCrit$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="bbotk" data-topic="OptimInstance" data-id="clear"><a href='../../bbotk/html/OptimInstance.html#method-OptimInstance-clear'><code>bbotk::OptimInstance$clear()</code></a></span></li>
<li><span class="pkg-link" data-pkg="bbotk" data-topic="OptimInstance" data-id="format"><a href='../../bbotk/html/OptimInstance.html#method-OptimInstance-format'><code>bbotk::OptimInstance$format()</code></a></span></li>
<li><span class="pkg-link" data-pkg="bbotk" data-topic="OptimInstance" data-id="print"><a href='../../bbotk/html/OptimInstance.html#method-OptimInstance-print'><code>bbotk::OptimInstance$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="bbotk" data-topic="OptimInstanceBatch" data-id="eval_batch"><a href='../../bbotk/html/OptimInstanceBatch.html#method-OptimInstanceBatch-eval_batch'><code>bbotk::OptimInstanceBatch$eval_batch()</code></a></span></li>
<li><span class="pkg-link" data-pkg="bbotk" data-topic="OptimInstanceBatch" data-id="objective_function"><a href='../../bbotk/html/OptimInstanceBatch.html#method-OptimInstanceBatch-objective_function'><code>bbotk::OptimInstanceBatch$objective_function()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3tuning" data-topic="TuningInstanceBatchSingleCrit" data-id="assign_result"><a href='../../mlr3tuning/html/TuningInstanceBatchSingleCrit.html#method-TuningInstanceBatchSingleCrit-assign_result'><code>mlr3tuning::TuningInstanceBatchSingleCrit$assign_result()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3tuning" data-topic="TuningInstanceBatchSingleCrit" data-id="initialize"><a href='../../mlr3tuning/html/TuningInstanceBatchSingleCrit.html#method-TuningInstanceBatchSingleCrit-initialize'><code>mlr3tuning::TuningInstanceBatchSingleCrit$initialize()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TuningInstanceSingleCrit-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TuningInstanceSingleCrit-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TuningInstanceSingleCrit$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
